backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pydantic==1.10.7
pymongo==4.5.0
torch==2.0.1
transformers==4.30.2
Pillow==9.5.0
python-multipart==0.0.6
pyvi==0.1.1